+++ src/math.ts
@@ -1,12 +1,12 @@
 import {extCompletion, extData} from './constants.js';
-import {getCompletions} from './util.js';
+import {getCompletions, isWikiparseLoaded} from './util.js';
 import type {StreamParser} from '@codemirror/language';
 
 let mathFetch: Promise<void> | undefined;
 
 export const math: StreamParser<object> = {
 	startState() {
-		if (typeof wikiparse === 'object') {
+		if (isWikiparseLoaded()) {
 			mathFetch ??= (async () => {
 				const data: string[] = await (await fetch(`${wikiparse.CDN}/data/ext/math.json`)).json();
 				extData['math'] = new Set(data);