+++ eslint.dist.mjs
@@ -5,6 +5,7 @@ export default [
 	{
 		rules: {
 			'es-x/no-resizable-and-growable-arraybuffers': 0,
+			'es-x/no-set-prototype-difference': 0,
 			'es-x/no-set-prototype-intersection': 0,
 		},
 	},